feat: Implementando estrutura controller para rota de professores
